Move code to Ryujinx.Common
