+++ Ryujinx.Graphics.Gpu/Image/Texture.cs
@@ -7,6 +7,7 @@ using Ryujinx.Graphics.Texture.Astc;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using Ryujinx.Common.Logging;
 
 namespace Ryujinx.Graphics.Gpu.Image
 {
@@ -245,7 +246,10 @@ namespace Ryujinx.Graphics.Gpu.Image
 
             if (!_context.Capabilities.SupportsAstcCompression && _info.FormatInfo.Format.IsAstc())
             {
-                if (!AstcDecoder.TryDecodeToRgba8P(
+                Logger.PrintInfo(LogClass.Gpu, "Decoding ASTC texture");
+                var watch = Stopwatch.StartNew();
+
+                if (!AstcDecoder.TryDecodeToRgba8(
                     data.ToArray(),
                     _info.FormatInfo.BlockWidth,
                     _info.FormatInfo.BlockHeight,
@@ -257,6 +261,10 @@ namespace Ryujinx.Graphics.Gpu.Image
                 {
                     // TODO: Error.
                 }
+
+                watch.Stop();
+
+                Logger.PrintInfo(LogClass.Gpu, $"Decoded texture in {watch.Elapsed.TotalSeconds} seconds");
 
                 data = decoded;
             }
